requirements.txt
Flask==3.0.*
flask-socketio==5.3.*
azure-cosmos==4.5.*